+++ src/sw.ts
@@ -1,5 +1,5 @@
 // @ts-expect-error https://dev.to/jacksteamdev/advanced-config-for-rpce-3966#dynamic-content-scripts
-import illustMetaInfo from './features/illust-meta-info?script';
+import content from './content?script';
 
 chrome.tabs.onUpdated.addListener((tabId, { status }, { url }) => {
   if (!url) {
@@ -15,7 +15,7 @@ chrome.tabs.onUpdated.addListener((tabId, { status }, { url }) => {
   chrome.scripting
     .executeScript({
       target: { tabId },
-      files: [illustMetaInfo as string],
+      files: [content as string],
     })
     .catch(console.error);
 });
