+++ src/vendors/setup-dayjs.ts
@@ -1,0 +1,7 @@
+import { extend, locale } from 'dayjs';
+import 'dayjs/locale/ja';
+import relativeTime from 'dayjs/plugin/relativeTime';
+
+locale('ja');
+
+extend(relativeTime);
